[bench] use ForceAttemptHTTP2 for http2
Go1.13しか使えないのでbuild constraintで分けておく
refs: #138
